requirements.txt
numpy
opencv-python
pillow
matplotlib
scikit-image
pywavelets
streamlit
torch
torchvision
torchaudio
jupyter
notebook
